requirements.txt
fastapi==0.85.1
uvicorn==0.19.0
python-dotenv==0.21.0
pydantic==1.10.2
passlib==1.7.4
pymongo==4.3.2
mongoengine==0.24.2
paho-mqtt==1.6.1
